pyproject.toml
[build-system]
requires = ["flit_core>=3.10.1,<4"]
build-backend = "flit_core.buildapi"

[project]
name = 'pt-repost'
version = '0'
requires-python = '~=3.10'
description = ''

dependencies = [
    'orjson~=3.10',
    'uuid-utils~=0.10',
    'rich~=13.9',
    'cryptography~=44.0',
    'xxhash~=3.5',
    'qbittorrent-api==2024.11.70',
    'click~=8.1.7',
    'sslog==0.0.0a51',
    'pydantic==2.10.1',
    'httpx~=0.28.0',
    'bencode2~=0.3.15',
    'tomli~=2.1.0',
    'yarl~=1.18',
    'guessit~=3.8',
    'six~=1.17',
    'psycopg[binary,pool]~=3.2.3',
    'regex~=2024.11',
    'rapidocr_onnxruntime~=1.4', # 识别硬字幕
    'pillow~=11.0',
    'durationpy',
    'fastapi',
    'uvicorn',
    'asyncpg',
    'jinja2'
]

[dependency-groups]
dev = [
    'types-regex~=2024.11',
    'types-six~=1.17',
    'mypy'
]

[project.scripts]
pt-repost = 'pt_repost.main:main'

[tool.pytest.ini_options]
cache_dir = '.venv/.cache/pytest'

[tool.mypy]
exclude = 'build,data'
mypy_path = '$MYPY_CONFIG_FILE_DIR'
python_version = '3.10'
disable_error_code = ["no-any-return"]
strict = true
ignore_missing_imports = true
check_untyped_defs = true

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100

[tool.ruff.format]

[tool.ruff.lint]
preview = true
extend-select = ['I', 'PLW1514']
unfixable = ['F401']
